Convert to upstart.
The init script would not run consistently for me on Ubuntu 12.04 —
it would hang on restart, allow multiple copies to be started, or claim
that it started OK while silently crashing.

Seems like this was probably because of `start-stop-daemon --background`,
which is warned against in its own man page.

Upstart is way easier.